requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
SQLAlchemy==2.0.29
asyncpg==0.29.0
psycopg[binary]==3.1.18
alembic==1.13.1
pydantic-settings==2.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
APScheduler==3.10.4
pydantic==2.7.1
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==4.1.0
python-dotenv==1.0.1
